move batching handlers to utils
